+++ table/example_test.go
@@ -179,37 +179,3 @@ func Example_alterTable() {
 		fmt.Printf("unexpected error: %v", err)
 	}
 }
-
-func Example_lazyTransaction() {
-	ctx := context.TODO()
-	db, err := ydb.Open(ctx, "grpcs://localhost:2135/local")
-	if err != nil {
-		fmt.Printf("failed connect: %v", err)
-		return
-	}
-	defer db.Close(ctx)
-	err = db.Table().Do(ctx,
-		func(ctx context.Context, s table.Session) (err error) {
-			// lazy open transaction on first execute query
-			tx, res, err := s.Execute(ctx, table.DefaultTxControl(), "SELECT 1", nil)
-			if err != nil {
-				return err // for auto-retry with driver
-			}
-			defer res.Close() // cleanup resources
-			if err = res.Err(); err != nil {
-				return err
-			}
-			// close transaction on last execute query
-			res, err = tx.WithCommit().Execute(ctx, "SELECT 2", nil)
-			if err != nil {
-				return err
-			}
-			defer res.Close()
-			return res.Err()
-		},
-		table.WithIdempotent(),
-	)
-	if err != nil {
-		fmt.Printf("unexpected error: %v", err)
-	}
-}